requirements.txt
gunicorn
google-cloud-bigquery
pandas
nltk
vaderSentiment
flask
